Refactor function to add notification to queue
